13619 refac store consumer handler (#514)
* #13619 updated dispatcher registration

* #13619 refactorings in flux area

* #13619 Incooperated IFluxDispatcher in IoC/Solution

Added the IFluxDispatcher interface to various classes and commands across the project for depenendecy injection of FluxDispatcher.cs.

* #13619 Refactor FluxConsumerBase into FluxConsumerExtensions (PART I)

* #13619 Fixed all FluxStoreBase and FluxStoreWithConsumerBase sub-classes

* #13619 Migrated to FluxConsumerExtensions.cs (Part II)

* #13619 Migrated to FluxConsumerExtensions.cs (Part IV)

* #13619 added a fallback handler in flux-consumer for compatibility reasons
AblageFaecherViewModel.cs was such fallback candidate. In case you need to subscribe to all current and future action verbs it can be error-prone to add the action verbs to the handler map, because then you have to keep track
with new action verbs of that store and add a mapping. the fallback handler reliefs you from that burdon. But keep in mind: It's just a fallback way for such scenario.

* #13619 Corrected typo in AppSpecificPropertiesExtensions.cs documentation.